drupal 8.9.15 -> 8.9.16
(Obtained from : https://www.drupal.org/project/drupal/releases/8.9.16) 

Maintenance and security release of the Drupal 8 series.

This release fixes a **security vulnerability**. Sites are **[urged to upgrade immediately]** after reading the notes below and the security announcement:

- [Drupal core - Moderately critical - Cross Site Scripting - SA-CORE-2021-003]

No other fixes are included.

Versions of Drupal 8 prior to 8.9.x are end-of-life and do not receive security coverage.

### Important update information

No changes have been made to the `.htaccess`, `web.config`, `robots.txt` or default `settings.php` files in this release, so upgrading custom versions of those files is not necessary if your site is already on the previous release.